bugzilla 110316 gmf-head 051207 Compiler warning count should be 0 - runtime diagram layer

+++ org.eclipse.gmf.runtime.draw2d.ui/src/org/eclipse/gmf/runtime/draw2d/ui/geometry/PointListUtilities.java
@@ -1515,7 +1515,7 @@ public class PointListUtilities {
 
 		long nPolyLength = length(mySegments);
 		long nCenterDistance =
-			(int) Math.round(
+			Math.round(
 				distanceAlong(mySegments, ptCenter)
 					* nPolyLength);
